requirements.txt
apache-beam[gcp]==2.55.0
tensorflow==2.15.0
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
matplotlib==3.8.4
shap==0.45.0
streamlit==1.36.0
confluent-kafka==2.5.0
kafka-python==2.0.2
google-cloud-storage==2.16.0
gcsfs==2024.6.1
pyarrow==16.1.0
